requirements.txt
#NOTE: virtual environment setup with Python 3.11
ibm-watsonx-orchestrate == 1.9.0
apify-client == 2.0.0
beautifulsoup4 == 4.13.4
bs4 == 0.0.2
requests == 2.32.4
selenium == 4.35.0
pillow == 11.3.0